Crate distance_util

Utility functionality useful when writing Rust tools for Distance.

The current functionality includes listing official levels, creating a level's leaderboard name string, and formatting a raw time or score obtained from the Steamworks API.

Structs

LevelNameTooLongError
The error returned by create_leaderboard_name_string when the resulting leaderboard name is too long.

NegativeScoreError
The error returned when a negative score is passed in to format_score.

Enums

LeaderboardGameMode
All Distance game modes that include leaderboards.

Constants

DEVELOPER_STEAM_IDS
Steam IDs of Distance's developers.

FOUNDING_MODDER_STEAM_IDS
Steam IDs of Founding Modders.

Functions

create_leaderboard_name_string
Creates a level's leaderboard name string, which is needed to get a level's leaderboard from the Steamworks API.

format_score
Returns a string representation of a raw "score".

official_level_leaderboard_names
Returns an iterator that yields the leaderboard name of all official levels for the given game mode.

official_level_names
Returns the level names of all official levels for the given game mode.