Function distance_util::create_leaderboard_name_string

pub fn create_leaderboard_name_string(
    level: &str, 
    game_mode: LeaderboardGameMode, 
    author_steam_id: Option<u64>
) -> Result<String, LevelNameTooLongError>

Creates a level's leaderboard name string, which is needed to get a level's leaderboard from the Steamworks API.

For official levels, level is the level's name. For workshop levels, level is the level's filename without the .bytes extension (which can be different from the level's title). author_steam_id must be given for workshop levels, and None for official levels.

Errors

The Steamworks API imposes a limit on the length of leaderboard names, so this function errors if a sufficiently long level name is given.