Function distance_util::official_level_leaderboard_names

pub fn official_level_leaderboard_names(
    mode: LeaderboardGameMode
) -> impl Iterator<Item = String> + Send

Returns an iterator that yields the leaderboard name of all official levels for the given game mode.

The returned name is used in the Steamworks API as a key to fetch a level's leaderboard.

Example

use distance_util::LeaderboardGameMode;

let mut names = distance_util::official_level_leaderboard_names(LeaderboardGameMode::Sprint);
assert_eq!(names.next().unwrap(), "Broken Symmetry_1_stable");