Function distance_util::format_score

pub fn format_score(
    score: i32,
    game_mode: LeaderboardGameMode
) -> Result<String, NegativeScoreError>

Returns a string representation of a raw “score”.

The Steamworks API returns a raw “score” as an i32 (milliseconds for Sprint and Challenge modes, and points for Stunt). This function can convert that into a String that matches how the time or score would look like in-game.

Times above 24 hours are correctly handled, compared to in-game, where the hour count wraps back around to zero.

Returns an error if score is negative.

Examples

use distance_util::LeaderboardGameMode;

let sprint_time = distance_util::format_score(83450, LeaderboardGameMode::Sprint).unwrap();
assert_eq!(sprint_time, "01:23.45");

let long_sprint_time = distance_util::format_score(17767890, LeaderboardGameMode::Sprint).unwrap();
assert_eq!(long_sprint_time, "04:56:07.89");

let stunt_time = distance_util::format_score(123_456, LeaderboardGameMode::Stunt).unwrap();
assert_eq!(stunt_time, "123,456 eV");